requirements.txt
# Core
streamlit>=1.36.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain/LangGraph
langchain>=0.1.0
langgraph>=0.1.0
langchain-core>=0.1.0
langsmith>=0.1.0

# LLM Providers (Groq is primary, others are fallback)
langchain-groq>=0.1.0
google-generativeai>=0.5.0
requests>=2.31.0

# Search
tavily-python>=0.3.0

# HTTP Client
httpx>=0.27.0

# Yahoo Finance
yfinance>=0.2.40

# A2A Server (for Researcher A2A mode)
fastapi>=0.115.0
uvicorn>=0.32.0

# MCP SDK (for MCP servers)
mcp>=1.0.0

# Sentiment Analysis
vaderSentiment>=3.3.2
